[fix] GetUniqueIdV2中rmtBufferVec.size()显式转换为u32，保持与BufferVecPack序列化一致

+++ src/legacy/unified_platform/resource/transport/ub_mem_transport.cc
@@ -818,7 +818,7 @@ std::vector<char> UbMemTransport::GetUniqueIdV2()
     binaryStream << type;
     binaryStream << notifyNum;
     binaryStream << bufferNum;
-    binaryStream << rmtBufferVec.size();
+    binaryStream << static_cast<u32>(rmtBufferVec.size());
     binaryStream << connNum;
  
     auto notifyUniqueIds = GetNotifyUniqueIds();
